requirements.txt
phonenumbers
opencage
folium
geopy
python-dotenv
python-dateutil
ttkbootstrap
pyperclip
requests
streamlit
streamlit-folium
pandas
kivy
buildozer
pyjnius